+++ lib/octopress-deploy.rb
@@ -34,11 +34,15 @@ module Octopress
     def self.pull(options={})
       options = merge_configs(options)
 
-      if !File.exists? options[:dir]
+      if Dir.exist?(options[:dir]) &&
+          !(Dir.entries(options[:dir]) - %w{. ..}).empty? &&
+          !options[:force]
+            puts "Pull failed. Directory #{options[:dir]} is not empty. Pass --force to overwrite."
+            abort
+      else
         FileUtils.mkdir_p options[:dir]
+        deployer(options).pull
       end
-
-      deployer(options).pull
     end
 
     def self.add_bucket(options={})
@@ -144,7 +148,9 @@ end
 
 Octopress::Docs.add({
   name:        "Octopress Deploy",
-  description: "An Octopress plugin for deploying static websites.",
+  gem:         "octopress-deploy",
+  version:     Octopress::Deploy::VERSION,
+  description: "Easily deploy any static site using S3, Git or Rsync.",
   path:        File.expand_path(File.join(File.dirname(__FILE__), "../")),
   source_url:  "https://github.com/octopress/deploy",
 })