+++ lib/octopress-deploy/commands.rb
@@ -15,6 +15,7 @@ module Octopress
           c.command(:pull) do |c|
             c.syntax "pull <DIR>"
             c.description "Pull down the published copy of your site into DIR"
+            c.option 'force', '--force', 'Overwrite existing files on pull'
             c.option "config_file", "--config FILE", "The path to your config file (default: _deploy.yml)"
 
             c.action do |args, options|
@@ -79,15 +80,15 @@ module Octopress
           end
 
           c.command(:'add-bucket') do |c|
-            c.syntax 'add-bucket [options]'
-            c.description "Add a new S3 bucket and configure it for static websites."
-            c.option 'bucket_name','--name NAME','Choose a bucket name. (Defaults: to bucket_name in config file)'
+            c.syntax 'add-bucket <NAME> [options]'
+            c.description "Add a new S3 bucket and configure it for static websites. Name defaults to bucket_name in config file"
             c.option 'region','--region REGION','Choose a region. (Defaults: to region in config file)'
             c.option 'index_page','--index PAGE','Specify an index page. (Default: index.html)'
             c.option 'error_page','--error PAGE','Specify an error page. (Default: 404.html)'
             c.option "config_file", "--config FILE", "The path to your config file (default: _deploy.yml)"
 
             c.action do |args, options|
+              options['bucket_name'] = args.first
               Octopress::Deploy.add_bucket(options)
             end
           end